+++ packages/compiler/src/extractor/extract.ts
@@ -20,6 +20,7 @@ import {
   componentHandler,
 } from "@kuma-ui/core/components/componentList";
 import { theme } from "@kuma-ui/sheet";
+import { stableStringify } from "../utils/stableStringify";
 
 export const extractProps = (
   componentName: (typeof componentList)[keyof typeof componentList],
@@ -34,19 +35,16 @@ export const extractProps = (
   // eslint-disable-next-line @typescript-eslint/no-explicit-any -- FIXME
   const componentProps: { [key: string]: any } = {};
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
   const variant = theme.getVariants(componentName);
   // eslint-disable-next-line @typescript-eslint/no-explicit-any -- FIXME
   const componentVariantProps: { [key: string]: any } = {
     ...(variant?.baseStyle as Record<string, string>),
   };
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call -- FIXME
   const defaultProps = componentDefaultProps(componentName);
 
   let isDefault = false;
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
   for (const [propName, propValue] of Object.entries({
     ...defaultProps,
     ...propsMap,
@@ -55,7 +53,6 @@ export const extractProps = (
       styledProps[propName.trim()] = propValue;
     } else if (isPseudoProps(propName.trim())) {
       pseudoProps[propName.trim()] = propValue;
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-call -- FIXME
     } else if (isComponentProps(componentName)(propName.trim())) {
       componentProps[propName.trim()] = propValue;
     } else if (propName.trim() === "variant") {
@@ -79,7 +76,6 @@ export const extractProps = (
     return;
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment, @typescript-eslint/no-unsafe-call -- FIXME
   const specificProps = componentHandler(componentName)(componentProps);
 
   // Every component internally uses the Box component.
@@ -91,20 +87,16 @@ export const extractProps = (
       }
     }
   }
-
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- FIXME
   const combinedProps = {
     ...componentVariantProps,
     ...specificProps,
     ...styledProps,
     ...pseudoProps,
   };
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
-  const key = generateKey(combinedProps);
+  const key = stableStringify(combinedProps);
   let generatedStyle = styleCache[key];
   // If the result isn't in the cache, generate it and save it to the cache
   if (!generatedStyle) {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
     generatedStyle = new StyleGenerator(combinedProps).getStyle();
     styleCache[key] = generatedStyle;
   }
@@ -157,19 +149,6 @@ export const extractProps = (
   return { css };
 };
 
-/**
- * Generates a unique key for props, aiding cache efficiency.
- * Incurs O(n log n) cost due to sorting, but it's acceptable given the
- * expensive nature of StyleGenerator's internals.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any -- FIXME
-const generateKey = (props: Record<string, any>) => {
-  return Object.entries(props)
-    .filter(([, value]) => value !== undefined)
-    .sort((a, b) => a[0].localeCompare(b[0]))
-    .map(([key, value]) => `${key}:${value}`)
-    .join("|");
-};
 const styleCache: {
   [key: string]: { className: string; css: string } | undefined;
 } = {};