+++ packages/core/src/components/Box/react/StaticBox.tsx
@@ -6,5 +6,5 @@ export const StaticBox: BoxComponent = ({
   children,
   IS_KUMA_DEFAULT,
   ...props
-// eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument -- FIXME
 }) => React.createElement(Component, props, children);