+++ packages/core/src/tests/vitest.setup.ts
@@ -1,0 +1,16 @@
+import type { TestingLibraryMatchers } from "@testing-library/jest-dom/matchers";
+import matchers from "@testing-library/jest-dom/matchers";
+import { cleanup } from "@testing-library/react";
+import { beforeEach, expect } from "vitest";
+
+declare module "vitest" {
+  interface Assertion<T>
+    extends jest.Matchers<void, T>,
+      TestingLibraryMatchers<T, void> {}
+}
+
+expect.extend(matchers);
+
+beforeEach(() => {
+  cleanup();
+});