+++ packages/core/src/theme.ts
@@ -5,11 +5,14 @@ import {
   flattenObject,
 } from "./utils/object";
 import { If, IsNever, Stringify, _String } from "./utils/types";
+import {
+  ResultThemeTokens,
+  SystemThemeTokens,
+  InputThemeTokens,
+} from "./themeTokens";
 import { componentList } from "./components/componentList";
 
-export type ThemeInput = {
-  colors?: NestedObject<string>;
-  breakpoints?: Record<string, string>;
+export type ThemeInput = InputThemeTokens & {
   components?: {
     [_ in keyof typeof componentList]?: {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any -- FIXME
@@ -21,47 +24,37 @@ export type ThemeInput = {
 };
 
 type ThemeResult<T extends ThemeInput> = Pretty<
-  ThemeColorResult<T> & ThemeComponentsResult<T> & ThemeBreakPointsResult<T>
+  ResultThemeTokens<Omit<T, "components">> & ThemeComponentsResult<T>
 >;
-
-type ThemeColorResult<T extends ThemeInput> = {
-  colors: T["colors"] extends NestedObject<string>
-    ? Pretty<FlattenObject<Pick<T, "colors">>>
-    : undefined;
-};
 
 type ThemeComponentsResult<T extends ThemeInput> = {
   components: T["components"];
 };
 
-type ThemeBreakPointsResult<T extends ThemeInput> = {
-  breakpoints: T["breakpoints"];
-};
-
- 
 export interface Theme {}
 
-type ThemeColors = Theme extends { colors: unknown } ? Theme["colors"] : never;
 type ThemeComponents = Theme extends { components: unknown }
   ? Theme["components"]
   : never;
-type ThemeBreakPoints = Theme extends { breakPoints: unknown }
-  ? Theme["breakPoints"]
-  : never;
-
 export type ThemeSystem = {
-  colors: If<IsNever<ThemeColors>, _String, Stringify<keyof ThemeColors>>;
   components: If<IsNever<ThemeComponents>, unknown, ThemeComponents>;
-  breakpoints: If<IsNever<ThemeBreakPoints>, unknown, ThemeBreakPoints>;
-};
+} & SystemThemeTokens;
 
 export function createTheme<const T extends ThemeInput>(
   theme: T
 ): ThemeResult<T> {
-  return {
-    colors: theme.colors ? flattenObject({ colors: theme.colors }) : undefined,
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- FIXME
+  const { components, ...tokens } = theme;
+  const resolvedTokens = {};
+  for (const key in tokens) {
+    // @ts-expect-error type
     // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- FIXME
-    components: theme.components,
-    breakpoints: theme.breakpoints,
+    resolvedTokens[key] = flattenObject({ [key]: tokens[key] });
+  }
+
+  return {
+    ...resolvedTokens,
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- FIXME
+    components: components,
   } as unknown as ThemeResult<T>;
 }