+++ packages/system/src/font.ts
@@ -1,17 +1,23 @@
 import { toCssUnit } from "./toCSS";
 import { FontKeys } from "./keys";
-import { ResponsiveStyle, CSSProperties } from "./types";
+import {
+  ResponsiveStyle,
+  CSSProperties,
+  AddProperty,
+  ThemeSystemType,
+} from "./types";
 import { applyResponsiveStyles } from "./responsive";
+import { theme } from "@kuma-ui/sheet";
 
-export type FontProps = Partial<
+export type FontProps<T extends ThemeSystemType = ThemeSystemType> = Partial<
   CSSProperties<"font"> &
-    CSSProperties<"fontFamily"> &
+    AddProperty<CSSProperties<"fontFamily">, T["fonts"]> &
     CSSProperties<"fontFeatureSettings"> &
     CSSProperties<"fontKerning"> &
     CSSProperties<"fontLanguageOverride"> &
     CSSProperties<"fontOpticalSizing"> &
     CSSProperties<"fontPalette"> &
-    CSSProperties<"fontSize", true> &
+    AddProperty<CSSProperties<"fontSize", true>, T["fontSizes"]> &
     CSSProperties<"fontSizeAdjust"> &
     CSSProperties<"fontStretch"> &
     CSSProperties<"fontStyle"> &
@@ -25,7 +31,7 @@ export type FontProps = Partial<
     CSSProperties<"fontVariantNumeric"> &
     CSSProperties<"fontVariantPosition"> &
     CSSProperties<"fontVariationSettings"> &
-    CSSProperties<"fontWeight", true>
+    AddProperty<CSSProperties<"fontWeight", true>, T["fontWeights"]>
 >;
 
 const fontMappings: Record<FontKeys, string> = {
@@ -61,9 +67,26 @@ export const font = (props: FontProps): ResponsiveStyle => {
     const cssValue = props[key as FontKeys];
     if (cssValue != undefined) {
       const property = fontMappings[key as FontKeys];
-      const converter = [fontMappings.fontSize].includes(property)
-        ? toCssUnit
-        : undefined;
+
+      const userTheme = theme.getUserTheme();
+      const converter = (value: string | number): string | number => {
+        if (property === "font-family") {
+          if (value in (userTheme.fonts ?? {})) {
+            return userTheme.fonts?.[value] as string;
+          }
+        } else if (property === "font-size") {
+          if (value in (userTheme.fontSizes ?? {})) {
+            return toCssUnit(userTheme.fontSizes?.[value] as string | number);
+          }
+          return toCssUnit(value);
+        } else if (property === "font-weight") {
+          if (value in (userTheme.fontWeights ?? {})) {
+            return userTheme.fontWeights?.[value] as string | number;
+          }
+        }
+        return value;
+      };
+
       const responsiveStyles = applyResponsiveStyles(
         property,
         cssValue,