+++ packages/system/src/space.test.ts
@@ -1,10 +1,21 @@
 import { space, SpaceProps } from "./space";
-import { describe, expect, test } from "vitest";
+import { beforeAll, describe, expect, test } from "vitest";
+import { theme } from "@kuma-ui/sheet";
 
 describe("space utility function", () => {
+  beforeAll(() => {
+    theme.setUserTheme({
+      spacings: {
+        sm: "1rem",
+      },
+    });
+  });
+
   // Arrange
   const testCases: Array<[SpaceProps, string, string]> = [
+    [{ margin: 8 }, "margin: 8px;", ""],
     [{ m: 8 }, "margin: 8px;", ""],
+    [{ marginTop: "8px" }, "margin-top: 8px;", ""],
     [{ mt: "8px" }, "margin-top: 8px;", ""],
     [{ mx: 8 }, "margin-left: 8px; margin-right: 8px;", ""],
     [
@@ -13,6 +24,7 @@ describe("space utility function", () => {
       "@media (min-width: 576px) { margin-left: 8px; margin-right: 8px; }",
     ],
     [{ m: 0 }, "margin: 0px;", ""],
+    [{ margin: "sm" }, "margin: 1rem;", ""],
   ];
 
   test.each(testCases)(