+++ packages/system/src/types.ts
@@ -1,14 +1,11 @@
 import * as CSS from "csstype";
 import { StyledKeyType } from "./keys";
+import { Tokens } from "@kuma-ui/sheet";
 
 type If<C extends boolean, T, F> = C extends true ? T : F;
 
 // eslint-disable-next-line @typescript-eslint/ban-types
-type _String = string & {};
-
-export type ThemeSystemType = {
-  colors: _String;
-};
+export type ThemeSystemType = Record<Tokens, string & {}>;
 
 // A type for non-undefined CSS property values
 export type CSSProperty = Exclude<
@@ -57,4 +54,8 @@ export type SystemStyle = {
     base: ResponsiveStyle["base"];
     responsive: ResponsiveStyle["media"];
   }[];
+};
+
+export type AddProperty<T, T2> = {
+  [Key in keyof T]: T[Key] | T2;
 };