+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @kuma-ui/vite
 
+## 1.0.9
+
+### Patch Changes
+
+- Updated dependencies [[`bc1a7f9`](https://github.com/kuma-ui/kuma-ui/commit/bc1a7f979f5539a4e2d1d2785dc2a893eee076eb), [`10f9811`](https://github.com/kuma-ui/kuma-ui/commit/10f98111d542f1fa45a6cda18ac872bec7b40910), [`066d043`](https://github.com/kuma-ui/kuma-ui/commit/066d04383d3f4e19fb0460b184e8aa9ed67ae41d)]:
+  - @kuma-ui/system@1.3.0
+  - @kuma-ui/sheet@1.1.0
+  - @kuma-ui/babel-plugin@1.1.5
+
+## 1.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`2329a1d`](https://github.com/kuma-ui/kuma-ui/commit/2329a1d9ea2fe4cf2d61cdf4ea87983c75b0d8fb)]:
+  - @kuma-ui/sheet@1.0.2
+  - @kuma-ui/system@1.2.1
+  - @kuma-ui/babel-plugin@1.1.4
+
+## 1.0.7
+
+### Patch Changes
+
+- [#275](https://github.com/kuma-ui/kuma-ui/pull/275) [`a0edbd5`](https://github.com/kuma-ui/kuma-ui/commit/a0edbd5b6ba0d7276524715f89221d7de5a403ad) Thanks [@kotarella1110](https://github.com/kotarella1110)! - fix(vite): fix hmr issue in vite v1.0.5 and later
+
+- Updated dependencies [[`022fdec`](https://github.com/kuma-ui/kuma-ui/commit/022fdecfd1c2494974e9965ddca31aebf03d1dc2), [`90a9e65`](https://github.com/kuma-ui/kuma-ui/commit/90a9e6504d9e11f164f2373a8968e8eef7130b9d)]:
+  - @kuma-ui/system@1.2.0
+  - @kuma-ui/babel-plugin@1.1.3
+
 ## 1.0.6
 
 ### Patch Changes